refactor(categories): move categories fetching logic to trpc procedure
The categories data fetching was moved from the SearchFilters component to a dedicated TRPC procedure to improve code organization and reusability. This change centralizes the data fetching logic and makes it available to other components through the TRPC router.
